Update the implementation for current D compilers
Build and language runtime:
- Build with current DMD/LDC releases. The D1 style operator overloads
  (opAdd, opSub, opMul, opDiv) were removed from the language, so the
  runtime object model now implements opBinary and forwards to them.
- Numbers are floating point and the lexer accepts a fractional part.
  Whole values are printed without a fractional tail so that counters and
  indexes keep reading as integers.
- Implement the and/or opcodes in the virtual machine, completing the
  instruction set listed in Table 4 and Appendix B.
- Both execution modes build from one source tree: `dub build` produces
  the standalone interpreter, `dub build --config=fcgi` the FastCGI one.

Runtime diagnostics:
- Type mismatches in arithmetic, division by zero, out of range and non
  integer array indexes, popping an empty array and incorrect built-in
  argument counts are reported as runtime errors with a message.
- Virtual machine stack pushes and pops are bounds checked against the
  128K slot buffer, and the buffer is registered as a garbage collector
  root since it is reached through a raw pointer.
- Equality compares values rather than type identity for strings,
  booleans and arrays.
- Unary minus is parsed and lowered to 0 - operand.
- The lexer and parser stay inside their buffers on truncated input, and
  string literals accept the \" \\ \n \t \r escapes.
- Errors reach the user as a single line instead of a stack trace.

Repository layout:
- Remove dub.sdl, which is the vendored fcgid library's own recipe and
  shadowed dub.json, and stop tracking .dub build artifacts.

+++ source/interlang.d
@@ -54,7 +54,7 @@ class IL{
 	}
 
 	/* load functions */
-	void load(int val){
+	void load(double val){
 		load(new RNumber(val));
 	}
 	void load(bool val){
